The package is ready to produce MT2Tree from the new version of ntuples (where the new tau identifications are stored. This version can be found at https://github.com/hbakhshi/NTupleProducer/commit/eec377339008d2139128059d7127f9a2184c080c )
In addition, a variable in the Makefile is added to control the hard-coded directories for the production (DATALOCALPATH). All the needed file for an independent production are added.

the eos version of the runManager is also added. It is a simple script to loop over files in eos and create bsub commands to submit jobs on the lxbatch

CAUTION : Old MT2Tau attributes might not be correct. There are some new variables to use.